fix: change cache buster step to use randoms (#307)
* Drop cache buster step

When we did builds on DockerHub or other external Docker providers, we
needed a way to bust the cache in the `Dockerfile` itself. However now
we do the builds on our own CI and simply push them to external
providers. We already disable the cache when we do this build on CI.
Plus builds start on fresh VMs without the old layers. Given this, let's
simplify and drop the flaky cache busting step.

* fix: leave cache busting steps commented out

* style: fix spacing

* Apply suggestions from code review

---------

Co-authored-by: Matthew R. Becker <beckermr@users.noreply.github.com>